Implemented PR feedback.
