+++ ql/instruments/Makefile.am
@@ -27,6 +27,7 @@ this_include_HEADERS = \
     dividendschedule.hpp \
     dividendvanillaoption.hpp \
     europeanoption.hpp \
+    fixedratebondforward.hpp \
     floatfloatswap.hpp \
     floatfloatswaption.hpp \
     forward.hpp \
@@ -86,6 +87,7 @@ cpp_files = \
 	dividendbarrieroption.cpp \
     dividendvanillaoption.cpp \
     europeanoption.cpp \
+    fixedratebondforward.cpp \
     floatfloatswap.cpp \
     floatfloatswaption.cpp \
     forward.cpp \
