+++ test-suite/bondforward.cpp
@@ -20,7 +20,6 @@
 #include "utilities.hpp"
 #include <ql/time/calendars/target.hpp>
 #include <ql/time/daycounters/actualactual.hpp>
-#include <ql/instruments/bonds/fixedratebond.hpp>
 #include <ql/instruments/bondforward.hpp>
 #include <ql/pricingengines/bond/discountingbondengine.hpp>
